solution/requirements.txt
python-dotenv==1.0.0
scipy==1.10.1
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
defusedxml==0.7.1
lxml==4.9.2
xmlschema==2.2.2
jupyterlab==3.6.1
networkx==3.0
py2neo==2021.2.3
apache-airflow==2.5.2